… java/src/main/java/com/adhoc/flight/utils/QueryUtils.java