# Now using Vector-based output

The purpose of this Pull Request is to replace boiterplate code with ready-made alternatives provided by Apache Arrow when it comes to manipulating `ValueVector`s (in this case, ones holding query results.) Since this repository should be the new user's best reference to look up to whenever creating an Arrow/Arrow Flight-compatible application, it makes sense to replace unnecessary verbose code with alternatives provided by Vector methods. Instead of merely thinking row-wise, this Pull Request implements a Vector-wise mindset, which is more appropriate when dealing with Arrow and Arrow Flight.

## Performance increase 
Because this PR uses a think-Vector-wise approach (as opposed to copying each Value again to place it in its corresponding row and column), there is a significant performance save. Here is a performance comparison for the query, `SELECT * FROM LocalFS1."dac-sample1.json"`:

### Old performance

```txt
Printing results to CONSOLE took: 112 ms.
The entire QUERY operation took: 733 ms.
```

### New performance

```txt
Printing results to CONSOLE took: 7 ms.
Saving results to FILE took: 8 ms.
The entire QUERY operation took: 679 ms.
```

## New functionality: save the binary data of the query results to a file
Whenever you run a new query, a new `VectorSchemaRoot` containing all query results is created. If you wish to save its bytes to a file, pass the CLI argument `-binpath <path>` or `--saveBinaryPath <path>` when running your query. Keep in mind that `path` is the relative path where you wish to save your file (or overwrite it from).